Implémentation de l'utilitaire de migration

+++ LSZ/LORENZ/Parametres.cs
@@ -8,10 +8,11 @@ namespace LORENZ
 {
     public static class Parametres
     {
+        private const string UserlogFileConst = "USERLOG.LZI";
         public static string LORENZPATH => Environment.CurrentDirectory;
         public static string ParamsDirectory => @"LZPARAMS";
-        public static string UserlogFile => $@"{ParamsDirectory}/USERLOG.LZI";
-        public static string LastAccessFile => $@"{ParamsDirectory}/LASTACSS.LZI";
+        public static string UserlogFile => Path.Combine(ParamsDirectory, UserlogFileConst);
+        public static string LastAccessFile => Path.Combine(ParamsDirectory, "LASTACSS.LZI");
         public static string CoinsRecordFile => $@"{ParamsDirectory}/COINSREC.LZI";
         public static string HelpFilePath => @"LZHELP.CHM";
         public static string ProductKeyFile => @"PRDCTKEY.LKI";
@@ -117,15 +118,32 @@ namespace LORENZ
                 {
                     if (!File.Exists(LastAccessFile))
                     {
-                        string LIDRetrieved = LireCleProduit().Item3;
-                        Display.PrintMessage("SUCCÈS: CLÉ DE PRODUIT VALIDE.", MessageState.Success);
-                        File.Delete(ProductKeyFile);
-                        // Write userinfos into USERLOG.LZI...
-                        Display.PrintMessage("Écriture des paramètres...", MessageState.Info);
-                        Directory.CreateDirectory(ParamsDirectory);
-                        EcrireParametres(LIDRetrieved);
-                        // Writing for first time LASTACSS.LZI...
-                        EcrireLastAccessFile(new FileInfo(UserlogFile).LastAccessTimeUtc);
+                        Console.Clear();
+                        Console.WriteLine("\nSi vous désirez migrer les paramètres d'une ancienne version de LORENZ, appuyez sur M.");
+                        Console.WriteLine("Si vous êtes au contraire un tout nouvel utilisateur, appuyez sur ENTRÉE pour poursuivre.");
+                        Console.WriteLine("Appuyez sur toute autre touche pour quitter.");
+                        ConsoleKey saisie = Console.ReadKey(true).Key;
+                        if (saisie == ConsoleKey.M)
+                        {
+                            MigrerParametres();
+                            continue;
+                        }
+                        else if (saisie == ConsoleKey.Enter)
+                        {
+                            string LIDRetrieved = LireCleProduit().Item3;
+                            Display.PrintMessage("SUCCÈS: CLÉ DE PRODUIT VALIDE.", MessageState.Success);
+                            File.Delete(ProductKeyFile);
+                            // Write userinfos into USERLOG.LZI...
+                            Display.PrintMessage("Écriture des paramètres...", MessageState.Info);
+                            Directory.CreateDirectory(ParamsDirectory);
+                            EcrireParametres(LIDRetrieved);
+                            // Writing for first time LASTACSS.LZI...
+                            EcrireLastAccessFile(new FileInfo(UserlogFile).LastAccessTimeUtc);
+                        }
+                        else
+                        {
+                            throw new LORENZException(ErrorCode.E0xFFF, false);
+                        }
                     }
                     else
                     {
@@ -188,27 +206,30 @@ namespace LORENZ
                     }
 
                     Console.Clear();
-                    if (essais == 1 && !File.Exists(ProductKeyFile))
+                    if (!File.Exists(ProductKeyFile))
                     {
-                        Display.PrintMessage("Bienvenue chez LORENZ ! En tant que nouvel utilisateur, voici quatre étapes", MessageState.Info);
-                        Display.PrintMessage("à suivre avant de pouvoir profiter de l'application :\n", MessageState.Info);
-                        Display.PrintMessage("1. Repérez le dossier de l'application \"LORENZSZ\". Pour ce faire, enfoncez", MessageState.Info);
-                        Display.PrintMessage("   les touches Windows + R de votre clavier et entrez %lorenzpath%.", MessageState.Info);
-                        Display.PrintMessage("2. Si vous avez installé \"CRYPTO\", vous trouverez un dossier du même nom", MessageState.Info);
-                        Display.PrintMessage("   dans ce répertoire. Entrez-y et exécutez le programme \"CRYPTO.exe\",", MessageState.Info);
-                        Display.PrintMessage("   puis, suivez les instructions qui s'afficheront.", MessageState.Info);
-                        Display.PrintMessage($"3. Insérez le fichier \"{ProductKeyFile}\" fourni par votre distributeur LORENZ", MessageState.Info);
-                        Display.PrintMessage("   dans le dossier du programme (c.-à-d. \"LORENZSZ\").", MessageState.Info);
-                        Display.PrintMessage("4. Appuyez sur n'importe quelle touche et nous vérifierons le reste.\n", MessageState.Info);
-                        Display.PrintMessage("NOTA : Si %lorenzpath% ne fonctionne pas, redémarrez votre ordinateur, puis réessayez.\n", MessageState.Warning);
-                        Display.PrintMessage("Appuyez sur ESC pour quitter.", MessageState.Info);
-                    }
-                    else if (!File.Exists(ProductKeyFile))
-                    {
-                        Display.PrintMessage($"Désolé! Nous n'avons pas détecté une clé \"{ProductKeyFile}\". Veuillez réessayer", MessageState.Warning);
-                        Display.PrintMessage("en vous assurant de l'avoir bien mis dans le dossier \"LORENZSZ\".", MessageState.Warning);
-                        Display.PrintMessage("Enfoncez les touches Windows + R et entrez %lorenzpath% pour le trouver.\n", MessageState.Warning);
-                        Display.PrintMessage("Appuyez sur ESC pour quitter.", MessageState.Warning);
+                        if (essais == 1)
+                        {
+                            Display.PrintMessage("Bienvenue chez LORENZ ! En tant que nouvel utilisateur, voici quatre étapes", MessageState.Info);
+                            Display.PrintMessage("à suivre avant de pouvoir profiter de l'application :\n", MessageState.Info);
+                            Display.PrintMessage("1. Repérez le dossier de l'application \"LORENZSZ\". Pour ce faire, enfoncez", MessageState.Info);
+                            Display.PrintMessage("   les touches Windows + R de votre clavier et entrez %lorenzpath%.", MessageState.Info);
+                            Display.PrintMessage("2. Si vous avez installé \"CRYPTO\", vous trouverez un dossier du même nom", MessageState.Info);
+                            Display.PrintMessage("   dans ce répertoire. Entrez-y et exécutez le programme \"CRYPTO.exe\",", MessageState.Info);
+                            Display.PrintMessage("   puis, suivez les instructions qui s'afficheront.", MessageState.Info);
+                            Display.PrintMessage($"3. Insérez le fichier \"{ProductKeyFile}\" fourni par votre distributeur LORENZ", MessageState.Info);
+                            Display.PrintMessage("   dans le dossier du programme (c.-à-d. \"LORENZSZ\").", MessageState.Info);
+                            Display.PrintMessage("4. Appuyez sur n'importe quelle touche et nous vérifierons le reste.\n", MessageState.Info);
+                            Display.PrintMessage("NOTA : Si %lorenzpath% ne fonctionne pas, redémarrez votre ordinateur, puis réessayez.\n", MessageState.Warning);
+                            Display.PrintMessage("Appuyez sur ESC pour quitter.", MessageState.Info);
+                        }
+                        else
+                        {
+                            Display.PrintMessage($"Désolé! Nous n'avons pas détecté une clé \"{ProductKeyFile}\". Veuillez réessayer", MessageState.Warning);
+                            Display.PrintMessage("en vous assurant de l'avoir bien mis dans le dossier \"LORENZSZ\".", MessageState.Warning);
+                            Display.PrintMessage("Enfoncez les touches Windows + R et entrez %lorenzpath% pour le trouver.\n", MessageState.Warning);
+                            Display.PrintMessage("Appuyez sur ESC pour quitter.", MessageState.Warning);
+                        }
                     }
                     else
                     {
@@ -220,7 +241,7 @@ namespace LORENZ
                     ConsoleKeyInfo saisie = Console.ReadKey(true);
                     while (saisie.Key == ConsoleKey.LeftWindows || saisie.Key == ConsoleKey.RightWindows)
                     {
-                        saisie = Console.ReadKey(true);
+                        saisie = Console.ReadKey(true);     // To override a key press on Windows Key
                     }
 
                     if (saisie.Key == ConsoleKey.Escape)
@@ -270,6 +291,65 @@ namespace LORENZ
                 throw new LORENZException(ErrorCode.E0x20, false);
             }
 
+        }
+
+        private static void MigrerParametres()
+        {
+            Console.Clear();
+            while (true)
+            {
+                Display.PrintMessage("Entrez le chemin d'accès complet vers le dossier LZPARAMS qui contient les anciens paramètres.");
+                Display.PrintMessage("Appuyez sur ESC pour quitter.");
+                Display.PrintMessage("NOTE : Si votre ancienne version correspond à LORENZ 2.0 ou antérieur, le chemin d'accès complet est :");
+                Display.PrintMessage("%localappdata%\\programs\\LORENZSZ\\LZPARAMS");
+                Console.Write("Chemin d'accès : ");
+                string pathOldParams = Extensions.SpecialInput();
+                if (pathOldParams == null)
+                {
+                    throw new LORENZException(ErrorCode.E0xFFF, false);
+                }
+
+                try
+                {
+                    // Test the specified directory path
+                    DirectoryInfo oldDirParams = new(pathOldParams);
+                    if (!oldDirParams.Exists)
+                    {
+                        oldDirParams.Create();
+                        oldDirParams.Delete();
+                    }
+                    else
+                    {
+                        string testdir = "testdir";
+                        oldDirParams.CreateSubdirectory(testdir);
+                        Directory.Delete(Path.Combine(pathOldParams, testdir));
+                    }
+
+                    // Check if USERLOG.LZI exists
+                    if (!new FileInfo(Path.Combine(pathOldParams, UserlogFileConst)).Exists)
+                    {
+                        Console.Clear();
+                        Display.PrintMessage($"Erreur ! Ce dossier ne contient pas de fichier {UserlogFileConst}.", MessageState.Failure);
+                        continue;
+                    }
+
+                    // Moving old directory params to new
+                    if (Directory.Exists(ParamsDirectory))
+                    {
+                        Directory.Delete(ParamsDirectory, true);
+                    }
+                    Directory.Move(pathOldParams, ParamsDirectory);
+
+                    // Rewriting LASTACSS.LZI
+                    EcrireLastAccessFile(new FileInfo(UserlogFile).LastAccessTimeUtc);
+                    return;
+                }
+                catch (Exception)
+                {
+                    Console.Clear();
+                    Display.PrintMessage("Erreur ! Le chemin d'accès spécifié est invalide.", MessageState.Failure);
+                }
+            }
         }
 
         private static void EcrireParametres(string LID)
